Moved Di vider.tsx out of /core and updated code patterns; added Create menu-specific styling to match prod
