+++ packages/manager/src/components/PrimaryNav/PrimaryNav.tsx
@@ -18,7 +18,7 @@ import TooltipIcon from 'src/assets/icons/get_help.svg';
 import Longview from 'src/assets/icons/longview.svg';
 import AkamaiLogo from 'src/assets/logo/akamai-logo.svg';
 import { BetaChip } from 'src/components/BetaChip/BetaChip';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import useAccountManagement from 'src/hooks/useAccountManagement';
 import useFlags from 'src/hooks/useFlags';
 import usePrefetch from 'src/hooks/usePreFetch';
