+++ packages/manager/src/features/Billing/BillingPanels/BillingSummary/BillingSummary.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Box } from 'src/components/Box';
 import { Button } from 'src/components/Button/Button';
 import { Currency } from 'src/components/Currency';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import Grid from '@mui/material/Unstable_Grid2';
 import PaymentDrawer from './PaymentDrawer';
 import PromoDialog from './PromoDialog';
