+++ packages/manager/src/features/Databases/DatabaseDetail/DatabaseSettings/DatabaseSettings.tsx
@@ -1,6 +1,6 @@
 import { Database } from '@linode/api-v4/lib/databases/types';
 import * as React from 'react';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import Paper from 'src/components/core/Paper';
 import { Typography } from 'src/components/Typography';
 import { useProfile } from 'src/queries/profile';
