+++ packages/manager/src/features/Kubernetes/KubeCheckoutBar/NodePoolSummary.tsx
@@ -1,7 +1,7 @@
 import Close from '@mui/icons-material/Close';
 import * as React from 'react';
 import { Box } from 'src/components/Box';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import { makeStyles } from '@mui/styles';
 import { Theme } from '@mui/material/styles';
 import { Typography } from 'src/components/Typography';
