+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeSettings/InterfaceSelect.tsx
@@ -3,7 +3,7 @@ import {
   InterfacePurpose,
 } from '@linode/api-v4/lib/linodes/types';
 import * as React from 'react';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import { makeStyles } from '@mui/styles';
 import { Theme, useTheme } from '@mui/material/styles';
 import Select, { Item } from 'src/components/EnhancedSelect/Select';
