+++ packages/manager/src/features/NodeBalancers/NodeBalancerConfigPanel.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import ActionsPanel from 'src/components/ActionsPanel';
 import { Button } from 'src/components/Button/Button';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import FormHelperText from 'src/components/core/FormHelperText';
 import Grid from '@mui/material/Unstable_Grid2';
 import { Link } from 'src/components/Link';
