+++ packages/manager/src/features/NotificationCenter/NotificationData/RenderEvent.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Box } from 'src/components/Box';
 import classNames from 'classnames';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import { HighlightedMarkdown } from 'src/components/HighlightedMarkdown/HighlightedMarkdown';
 import { Typography } from 'src/components/Typography';
 import useEventInfo from './useEventInfo';
