+++ packages/manager/src/features/Profile/AuthenticationSettings/AuthenticationSettings.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import { Link } from 'src/components/Link';
 import Paper from 'src/components/core/Paper';
 import TrustedDevices from './TrustedDevices';
