+++ packages/manager/src/features/Profile/AuthenticationSettings/TPAProviders.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import AkamaiWaveOnlyIcon from 'src/assets/icons/providers/akamai-logo-rgb-waveOnly.svg';
 import { Box } from 'src/components/Box';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import EnabledIcon from 'src/assets/icons/checkmark-enabled.svg';
 import ExternalLink from 'src/components/ExternalLink';
 import GitHubIcon from 'src/assets/icons/providers/github-logo.svg';
