+++ packages/manager/src/features/Support/Hively.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { DateTime } from 'luxon';
 import { parseAPIDate } from 'src/utilities/date';
-import Divider from 'src/components/core/Divider';
+import Divider from 'src/components/Divider';
 import { makeStyles } from '@mui/styles';
 import { Theme } from '@mui/material/styles';
 import { Typography } from 'src/components/Typography';
